backend/foodgram/requirements.txt
django-colorfield==0.7.2
django-cors-headers==3.13.0
Pillow==8.3.1
asgiref==3.2.10
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
webcolors==1.11.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1